requirements.txt
appdirs==1.4.3
backports.shutil-get-terminal-size==1.0.0
crypto==1.4.1
decorator==4.0.11
Django==1.9
django-braces==1.11.0
django-cors-middleware==1.3.1
django-filter==1.0.2
django-oauth-toolkit==0.12.0
django-oauth2-provider==0.2.6.1
django-rest-auth==0.9.1
djangorestframework==3.5.4
enum34==1.1.6
ipdb==0.10.3
ipython==5.3.0
ipython-genutils==0.2.0
Naked==0.1.31
oauthlib==2.0.1
olefile==0.44
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.14
psycopg2==2.7
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
python-http-client==2.2.1
PyYAML==3.12
reportlab==3.4.0
requests==2.14.2
scandir==1.5
sendgrid==3.6.5
sendgrid-django==4.0.4
shellescape==3.4.1
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
